refactor(client): Decompose LeftPanel into custom hook and smaller UI components
