+++ aiblog-client/src/components/LeftPanel/ErrorState.jsx
@@ -1,0 +1,9 @@
+import { Alert } from '@mui/material';
+
+export function ErrorState({ message }) {
+  return (
+    <Alert severity="error" sx={{ m: 2 }}>
+      {message}
+    </Alert>
+  );
+}
