+++ aiblog-client/src/components/LeftPanel/LoadingState.jsx
@@ -1,0 +1,17 @@
+import { Box, CircularProgress, Typography } from '@mui/material';
+
+export function LoadingState({ message = 'Loading data...' }) {
+  return (
+    <Box
+      sx={{
+        display: 'flex',
+        justifyContent: 'center',
+        alignItems: 'center',
+        height: '200px',
+      }}
+    >
+      <CircularProgress />
+      <Typography sx={{ ml: 2 }}>{message}</Typography>
+    </Box>
+  );
+}
